Product Details

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\CategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Str;
 
@@ -95,6 +96,11 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::findOrFail($id);
+        $subCategory = SubCategory::where('category_id',$category->id)->count();
+        if($subCategory > 0){
+            return response(['status' =>'error','message'=>'This Items Contains SubCategory,For Delete This Item You 
+            Have To Delete The SubCategory Item Firts!']);
+        }
         $category->delete();
         return response(['status'=>'success','Deleted Successfully']);
     }
